+++ scripts/helpers/spell-preparation.mjs
@@ -1,132 +1,13 @@
-import { CANTRIP_RULES, ENFORCEMENT_BEHAVIOR, FLAGS, MODULE, SETTINGS } from '../constants.mjs';
+import { CANTRIP_RULES, CLASS_IDENTIFIERS, ENFORCEMENT_BEHAVIOR, FLAGS, MODULE, SETTINGS } from '../constants.mjs';
 import { log } from '../logger.mjs';
+import { CantripManager } from './cantrip-manager.mjs';
 import * as genericUtils from './generic-utils.mjs';
+import { RitualManager } from './ritual-manager.mjs';
 import * as formattingUtils from './spell-formatting.mjs';
 import { WizardSpellbookManager } from './wizard-spellbook.mjs';
 
 /**
- * Save prepared spells for an actor
- * @param {Actor5e} actor - The actor to save spells for
- * @param {Object} spellData - Object of spell data with preparation info
- * @returns {Promise<void>}
- */
-export async function saveActorPreparedSpells(actor, spellData) {
-  try {
-    log(3, `Saving prepared spells for ${actor.name}`);
-    const cantripChanges = { added: [], removed: [], hasChanges: false };
-    const preparedUuids = Object.entries(spellData)
-      .filter(([_uuid, data]) => data.isPrepared)
-      .map(([uuid]) => uuid);
-
-    await actor.setFlag(MODULE.ID, FLAGS.PREPARED_SPELLS, preparedUuids);
-    log(3, `Saved ${preparedUuids.length} prepared spells to actor flags`);
-
-    const spellIdsToRemove = [];
-    const spellsToUpdate = [];
-    const spellsToCreate = [];
-
-    const isWizard = genericUtils.isWizard(actor);
-    const wizardManager = isWizard ? new WizardSpellbookManager(actor) : null;
-    const ritualCastingEnabled = wizardManager ? actor.getFlag(MODULE.ID, 'wizardRitualCasting') !== false : false;
-
-    for (const [uuid, data] of Object.entries(spellData)) {
-      if (data.isAlwaysPrepared) continue;
-      const isRitual = data.isRitual || false;
-      const existingSpell = actor.items.find((i) => i.type === 'spell' && (i.flags?.core?.sourceId === uuid || i.uuid === uuid));
-
-      if (!data.isPrepared) {
-        if (data.wasPrepared && existingSpell) {
-          if (isRitual && isWizard && ritualCastingEnabled) {
-            spellsToUpdate.push({
-              '_id': existingSpell.id,
-              'system.preparation.mode': 'ritual',
-              'system.preparation.prepared': false
-            });
-          } else if (existingSpell.system.preparation?.mode === 'prepared' && !existingSpell.system.preparation?.alwaysPrepared) {
-            spellIdsToRemove.push(existingSpell.id);
-            if (existingSpell.system.level === 0) {
-              cantripChanges.removed.push({ name: existingSpell.name, uuid: uuid });
-              cantripChanges.hasChanges = true;
-            }
-          }
-        } else if (isRitual && isWizard && ritualCastingEnabled && !existingSpell) {
-          try {
-            const sourceSpell = await fromUuid(uuid);
-            if (sourceSpell) {
-              const newSpellData = sourceSpell.toObject();
-              if (!newSpellData.system.preparation) newSpellData.system.preparation = {};
-              newSpellData.system.preparation.mode = 'ritual';
-              newSpellData.system.preparation.prepared = false;
-              newSpellData.flags = newSpellData.flags || {};
-              newSpellData.flags.core = newSpellData.flags.core || {};
-              newSpellData.flags.core.sourceId = uuid;
-              spellsToCreate.push(newSpellData);
-            }
-          } catch (error) {
-            log(1, `Error fetching ritual spell ${uuid}:`, error);
-          }
-        }
-      } else {
-        if (existingSpell) {
-          if (!existingSpell.system.preparation?.prepared || existingSpell.system.preparation?.mode !== 'prepared') {
-            spellsToUpdate.push({
-              '_id': existingSpell.id,
-              'system.preparation.mode': 'prepared',
-              'system.preparation.prepared': true
-            });
-          }
-        } else {
-          try {
-            const sourceSpell = await fromUuid(uuid);
-            if (sourceSpell) {
-              const newSpellData = sourceSpell.toObject();
-              if (!newSpellData.system.preparation) newSpellData.system.preparation = {};
-              newSpellData.system.preparation.mode = 'prepared';
-              newSpellData.system.preparation.prepared = true;
-              newSpellData.flags = newSpellData.flags || {};
-              newSpellData.flags.core = newSpellData.flags.core || {};
-              newSpellData.flags.core.sourceId = uuid;
-              spellsToCreate.push(newSpellData);
-
-              if (sourceSpell.system.level === 0) {
-                cantripChanges.added.push({ name: sourceSpell.name, uuid });
-                cantripChanges.hasChanges = true;
-              }
-            }
-          } catch (error) {
-            log(1, `Error fetching spell ${uuid}:`, error);
-          }
-        }
-      }
-    }
-
-    if (spellIdsToRemove.length > 0) {
-      log(3, `Removing ${spellIdsToRemove.length} spells from actor`);
-      await actor.deleteEmbeddedDocuments('Item', spellIdsToRemove);
-    }
-
-    if (spellsToUpdate.length > 0) {
-      log(3, `Updating ${spellsToUpdate.length} spells on actor`);
-      await actor.updateEmbeddedDocuments('Item', spellsToUpdate);
-    }
-
-    if (spellsToCreate.length > 0) {
-      log(3, `Creating ${spellsToCreate.length} spells on actor`);
-      await actor.createEmbeddedDocuments('Item', spellsToCreate);
-    }
-
-    if (cantripChanges.hasChanges) {
-      const spellManager = new SpellManager(actor);
-      await spellManager.notifyGMOfCantripChanges(cantripChanges);
-    }
-  } catch (error) {
-    log(1, `Error saving prepared spells for ${actor?.name || 'unknown actor'}:`, error);
-    ui.notifications.error(game.i18n.localize('SPELLBOOK.Error.SavingFailed'));
-  }
-}
-
-/**
- * Manages spell preparation and cantrip functionality
+ * Manages spell preparation and related functionality
  */
 export class SpellManager {
   /**
@@ -142,6 +23,10 @@ export class SpellManager {
     this.isWizard = genericUtils.isWizard(actor);
     this._wizardSpellbookCache = null;
     this._wizardManager = null;
+
+    // Initialize sub-managers
+    this.cantripManager = new CantripManager(actor, this);
+    this.ritualManager = new RitualManager(actor);
   }
 
   /**
@@ -167,16 +52,18 @@ export class SpellManager {
     }
 
     const classLevel = this.classItem.system.levels || this.actor.system.details.level;
-    switch (this.classItem.name.toLowerCase()) {
-      case 'bard':
-      case 'cleric':
-      case 'druid':
-      case 'sorcerer':
-      case 'warlock':
-      case 'wizard':
+    const className = this.classItem.name.toLowerCase();
+
+    switch (className) {
+      case CLASS_IDENTIFIERS.BARD:
+      case CLASS_IDENTIFIERS.CLERIC:
+      case CLASS_IDENTIFIERS.DRUID:
+      case CLASS_IDENTIFIERS.SORCERER:
+      case CLASS_IDENTIFIERS.WARLOCK:
+      case CLASS_IDENTIFIERS.WIZARD:
         return Math.min(4, Math.max(3, Math.floor(classLevel / 4) + 2));
-      case 'ranger':
-      case 'artificer':
+      case CLASS_IDENTIFIERS.RANGER:
+      case CLASS_IDENTIFIERS.ARTIFICER:
         return Math.min(3, Math.max(2, Math.floor(classLevel / 6) + 1));
       default:
         return 0;
@@ -219,99 +106,7 @@ export class SpellManager {
    * @returns {Object} Status object with allowed and message properties
    */
   canChangeCantripStatus(spell, isChecked, isLevelUp, isLongRest, uiCantripCount = null) {
-    if (spell.system.level !== 0) return { allowed: true };
-
-    const spellName = spell.name || 'unknown cantrip';
-    const { rules, behavior } = this.settings;
-
-    if (behavior === ENFORCEMENT_BEHAVIOR.UNENFORCED || behavior === ENFORCEMENT_BEHAVIOR.NOTIFY_GM) {
-      if (behavior === ENFORCEMENT_BEHAVIOR.NOTIFY_GM && isChecked) {
-        const currentCount = uiCantripCount !== null ? uiCantripCount : this.getCurrentCount();
-        if (currentCount >= this.maxCantrips) {
-          ui.notifications.warn(game.i18n.localize('SPELLBOOK.Cantrips.MaximumReached'));
-        }
-      }
-      return { allowed: true };
-    }
-
-    if (isChecked) {
-      const currentCount = uiCantripCount !== null ? uiCantripCount : this.getCurrentCount();
-      if (currentCount >= this.maxCantrips) {
-        return { allowed: false, message: 'SPELLBOOK.Cantrips.MaximumReached' };
-      }
-    }
-
-    const trackingData = this._getSwapTrackingData(isLevelUp, isLongRest);
-
-    if (rules === CANTRIP_RULES.MODERN_LONG_REST) {
-      if (!this.isWizard) {
-        return { allowed: false, message: 'SPELLBOOK.Cantrips.WizardRuleOnly' };
-      }
-      if (!isLongRest && !isChecked) {
-        return { allowed: false, message: 'SPELLBOOK.Cantrips.LockedOutsideLongRest' };
-      }
-      if (isLongRest) {
-        if (!isChecked && trackingData.hasUnlearned && trackingData.unlearned !== spell.uuid && trackingData.originalChecked.includes(spell.uuid)) {
-          return { allowed: false, message: 'SPELLBOOK.Cantrips.OnlyOneSwap' };
-        }
-        if (isChecked && trackingData.hasLearned && trackingData.learned !== spell.uuid && !trackingData.originalChecked.includes(spell.uuid)) {
-          return { allowed: false, message: 'SPELLBOOK.Cantrips.OnlyOneSwap' };
-        }
-        if (isChecked && !trackingData.hasUnlearned && !trackingData.originalChecked.includes(spell.uuid)) {
-          return { allowed: false, message: 'SPELLBOOK.Cantrips.MustUnlearnFirst' };
-        }
-      }
-    } else if (rules === CANTRIP_RULES.MODERN_LEVEL_UP) {
-      if (!isLevelUp && !isChecked) {
-        return { allowed: false, message: 'SPELLBOOK.Cantrips.LockedOutsideLevelUp' };
-      }
-      if (isLevelUp) {
-        if (!isChecked && trackingData.hasUnlearned && trackingData.unlearned !== spell.uuid && trackingData.originalChecked.includes(spell.uuid)) {
-          return { allowed: false, message: 'SPELLBOOK.Cantrips.OnlyOneSwap' };
-        }
-        if (isChecked && trackingData.hasLearned && trackingData.learned !== spell.uuid && !trackingData.originalChecked.includes(spell.uuid)) {
-          return { allowed: false, message: 'SPELLBOOK.Cantrips.OnlyOneSwap' };
-        }
-        if (isChecked && !trackingData.hasUnlearned && !trackingData.originalChecked.includes(spell.uuid)) {
-          return { allowed: false, message: 'SPELLBOOK.Cantrips.MustUnlearnFirst' };
-        }
-      }
-    } else if (rules === CANTRIP_RULES.LEGACY) {
-      if (!isChecked) {
-        return { allowed: false, message: 'SPELLBOOK.Cantrips.LockedLegacy' };
-      }
-    }
-
-    return { allowed: true };
-  }
-
-  /**
-   * Get the current swap tracking data
-   * @param {boolean} isLevelUp - Whether this is a level-up context
-   * @param {boolean} isLongRest - Whether this is a long rest context
-   * @returns {Object} Tracking data
-   * @private
-   */
-  _getSwapTrackingData(isLevelUp, isLongRest) {
-    let flagName;
-    if (isLevelUp) {
-      flagName = FLAGS.CANTRIP_SWAP_TRACKING + '.levelUp';
-    } else if (isLongRest) {
-      flagName = FLAGS.CANTRIP_SWAP_TRACKING + '.longRest';
-    } else {
-      return { hasUnlearned: false, unlearned: null, hasLearned: false, learned: null, originalChecked: [] };
-    }
-
-    const data = this.actor.getFlag(MODULE.ID, flagName);
-    return (
-      data || {
-        hasUnlearned: false,
-        unlearned: null,
-        hasLearned: false,
-        learned: null,
-        originalChecked: []
-      }
-    );
+    return this.cantripManager.canChangeCantripStatus(spell, isChecked, isLevelUp, isLongRest, uiCantripCount);
   }
 
   /**
@@ -322,74 +117,7 @@ export class SpellManager {
    * @param {boolean} isLongRest - Whether this is during a long rest
    */
   trackCantripChange(spell, isChecked, isLevelUp, isLongRest) {
-    if (spell.system.level !== 0) return;
-
-    const { rules } = this.settings;
-    const spellUuid = genericUtils.getSpellUuid(spell);
-
-    if (!isLevelUp && !isLongRest) return;
-    if (rules === CANTRIP_RULES.LEGACY) return;
-    if (rules === CANTRIP_RULES.MODERN_LONG_REST && !this.isWizard) return;
-
-    const flagName = isLevelUp ? FLAGS.CANTRIP_SWAP_TRACKING + '.levelUp' : FLAGS.CANTRIP_SWAP_TRACKING + '.longRest';
-    let tracking = this.actor.getFlag(MODULE.ID, flagName);
-
-    if (!tracking) {
-      const preparedCantrips = this.actor.items.filter((i) => i.type === 'spell' && i.system.level === 0 && i.system.preparation?.prepared).map((i) => genericUtils.getSpellUuid(i));
-
-      tracking = {
-        hasUnlearned: false,
-        unlearned: null,
-        hasLearned: false,
-        learned: null,
-        originalChecked: preparedCantrips
-      };
-      this.actor.setFlag(MODULE.ID, flagName, tracking);
-    }
-
-    if (!isChecked && tracking.originalChecked.includes(spellUuid)) {
-      if (tracking.unlearned === spellUuid) {
-        tracking.hasUnlearned = false;
-        tracking.unlearned = null;
-      } else {
-        tracking.hasUnlearned = true;
-        tracking.unlearned = spellUuid;
-      }
-    } else if (isChecked && !tracking.originalChecked.includes(spellUuid)) {
-      if (tracking.learned === spellUuid) {
-        tracking.hasLearned = false;
-        tracking.learned = null;
-      } else {
-        tracking.hasLearned = true;
-        tracking.learned = spellUuid;
-      }
-    } else if (!isChecked && tracking.learned === spellUuid) {
-      tracking.hasLearned = false;
-      tracking.learned = null;
-    } else if (isChecked && tracking.unlearned === spellUuid) {
-      tracking.hasUnlearned = false;
-      tracking.unlearned = null;
-    }
-
-    this.actor.setFlag(MODULE.ID, flagName, tracking);
-  }
-
-  /**
-   * Complete the cantrip swap process and reset tracking
-   * @param {boolean} isLevelUp - Whether this is completing a level-up swap
-   * @returns {Promise<boolean>} Success status
-   */
-  async completeCantripSwap(isLevelUp) {
-    const flagName = isLevelUp ? FLAGS.CANTRIP_SWAP_TRACKING + '.levelUp' : FLAGS.CANTRIP_SWAP_TRACKING + '.longRest';
-    await this.actor.unsetFlag(MODULE.ID, flagName);
-
-    if (isLevelUp) {
-      const currentLevel = this.actor.system.details.level;
-      const currentMax = this.getMaxAllowed();
-      await this.actor.setFlag(MODULE.ID, FLAGS.PREVIOUS_LEVEL, currentLevel);
-      await this.actor.setFlag(MODULE.ID, FLAGS.PREVIOUS_CANTRIP_MAX, currentMax);
-    }
-    return true;
+    this.cantripManager.trackCantripChange(spell, isChecked, isLevelUp, isLongRest);
   }
 
   /**
@@ -400,87 +128,7 @@ export class SpellManager {
    * @param {number} currentCount - Current count of prepared cantrips
    */
   lockCantripCheckboxes(cantripItems, isLevelUp, isLongRest, currentCount) {
-    const { rules, behavior } = this.settings;
-    const isAtMax = currentCount >= this.maxCantrips;
-    const trackingData = this._getSwapTrackingData(isLevelUp, isLongRest);
-
-    for (const item of cantripItems) {
-      const checkbox = item.querySelector('dnd5e-checkbox');
-      if (!checkbox) continue;
-
-      if (item.querySelector('.tag.always-prepared') || item.querySelector('.tag.granted')) continue;
-
-      const isChecked = checkbox.checked;
-      const uuid = checkbox.dataset.uuid;
-
-      checkbox.disabled = false;
-      delete checkbox.dataset.tooltip;
-      item.classList.remove('cantrip-locked');
-
-      if (isAtMax && !isChecked) {
-        checkbox.disabled = true;
-        checkbox.dataset.tooltip = game.i18n.localize('SPELLBOOK.Cantrips.MaximumReached');
-        item.classList.add('cantrip-locked');
-        continue;
-      }
-
-      if (behavior !== ENFORCEMENT_BEHAVIOR.ENFORCED) continue;
-
-      switch (rules) {
-        case CANTRIP_RULES.LEGACY:
-          if (isChecked) {
-            checkbox.disabled = true;
-            checkbox.dataset.tooltip = game.i18n.localize('SPELLBOOK.Cantrips.LockedLegacy');
-            item.classList.add('cantrip-locked');
-          }
-          break;
-
-        case CANTRIP_RULES.MODERN_LEVEL_UP:
-          if (isLevelUp) {
-            if (trackingData.hasUnlearned && uuid !== trackingData.unlearned && isChecked) {
-              checkbox.disabled = true;
-              checkbox.dataset.tooltip = game.i18n.localize('SPELLBOOK.Cantrips.OnlyOneSwap');
-              item.classList.add('cantrip-locked');
-            }
-            if (trackingData.hasLearned && uuid !== trackingData.learned && !isChecked) {
-              checkbox.disabled = true;
-              checkbox.dataset.tooltip = game.i18n.localize('SPELLBOOK.Cantrips.OnlyOneSwap');
-              item.classList.add('cantrip-locked');
-            }
-          } else if (isChecked) {
-            checkbox.disabled = true;
-            checkbox.dataset.tooltip = game.i18n.localize('SPELLBOOK.Cantrips.LockedOutsideLevelUp');
-            item.classList.add('cantrip-locked');
-          }
-          break;
-
-        case CANTRIP_RULES.MODERN_LONG_REST:
-          if (!this.isWizard) {
-            checkbox.disabled = true;
-            checkbox.dataset.tooltip = game.i18n.localize('SPELLBOOK.Cantrips.WizardRuleOnly');
-            item.classList.add('cantrip-locked');
-            continue;
-          }
-
-          if (isLongRest) {
-            if (trackingData.hasUnlearned && uuid !== trackingData.unlearned && isChecked) {
-              checkbox.disabled = true;
-              checkbox.dataset.tooltip = game.i18n.localize('SPELLBOOK.Cantrips.OnlyOneSwap');
-              item.classList.add('cantrip-locked');
-            }
-            if (trackingData.hasLearned && uuid !== trackingData.learned && !isChecked) {
-              checkbox.disabled = true;
-              checkbox.dataset.tooltip = game.i18n.localize('SPELLBOOK.Cantrips.OnlyOneSwap');
-              item.classList.add('cantrip-locked');
-            }
-          } else if (isChecked) {
-            checkbox.disabled = true;
-            checkbox.dataset.tooltip = game.i18n.localize('SPELLBOOK.Cantrips.LockedOutsideLongRest');
-            item.classList.add('cantrip-locked');
-          }
-          break;
-      }
-    }
+    this.cantripManager.lockCantripCheckboxes(cantripItems, isLevelUp, isLongRest, currentCount);
   }
 
   /**
@@ -489,31 +137,7 @@ export class SpellManager {
    * @returns {Promise<void>}
    */
   async notifyGMOfCantripChanges(changes) {
-    if (changes.added.length === 0 && changes.removed.length === 0) return;
-    if (this.settings.behavior !== ENFORCEMENT_BEHAVIOR.NOTIFY_GM) return;
-
-    const currentCantrips = this.actor.items.filter((i) => i.type === 'spell' && i.system.level === 0 && i.system.preparation?.prepared && !i.system.preparation?.alwaysPrepared).map((i) => i.name);
-
-    const originalCantripsSet = new Set(currentCantrips);
-    for (const { name } of changes.removed) originalCantripsSet.add(name);
-    for (const { name } of changes.added) originalCantripsSet.delete(name);
-    const originalCantrips = Array.from(originalCantripsSet).sort();
-
-    const newCantripsSet = new Set(originalCantrips);
-    for (const { name } of changes.removed) newCantripsSet.delete(name);
-    for (const { name } of changes.added) newCantripsSet.add(name);
-    const newCantrips = Array.from(newCantripsSet).sort();
-
-    let content = `<h3>${game.i18n.format('SPELLBOOK.Cantrips.ChangeNotification', { name: this.actor.name })}</h3>`;
-    if (originalCantrips.length > 0) content += `<p><strong>Original Cantrips:</strong> ${originalCantrips.join(', ')}</p>`;
-    if (changes.removed.length > 0) content += `<p><strong>${game.i18n.localize('SPELLBOOK.Cantrips.Removed')}:</strong> ${changes.removed.map((c) => c.name).join(', ')}</p>`;
-    if (changes.added.length > 0) content += `<p><strong>${game.i18n.localize('SPELLBOOK.Cantrips.Added')}:</strong> ${changes.added.map((c) => c.name).join(', ')}</p>`;
-    if (newCantrips.length > 0) content += `<p><strong>New Cantrips:</strong> ${newCantrips.join(', ')}</p>`;
-
-    ChatMessage.create({
-      content: content,
-      whisper: game.users.filter((u) => u.isGM).map((u) => u.id)
-    });
+    return this.cantripManager.notifyGMOfCantripChanges(changes);
   }
 
   /**
@@ -562,16 +186,7 @@ export class SpellManager {
    * @returns {boolean} Whether a level-up cantrip change is detected
    */
   checkForLevelUp() {
-    try {
-      const previousLevel = this.actor.getFlag(MODULE.ID, FLAGS.PREVIOUS_LEVEL) || 0;
-      const previousMax = this.actor.getFlag(MODULE.ID, FLAGS.PREVIOUS_CANTRIP_MAX) || 0;
-      const currentLevel = this.actor.system.details.level;
-      const currentMax = this.getMaxAllowed();
-      return (currentLevel > previousLevel || currentMax > previousMax) && previousLevel > 0;
-    } catch (error) {
-      log(1, 'Error checking for cantrip level up:', error);
-      return false;
-    }
+    return this.cantripManager.checkForLevelUp();
   }
 
   /**
@@ -579,11 +194,7 @@ export class SpellManager {
    * @returns {boolean} Whether cantrips can be changed
    */
   canBeLeveledUp() {
-    const previousLevel = this.actor.getFlag(MODULE.ID, FLAGS.PREVIOUS_LEVEL) || 0;
-    const previousMax = this.actor.getFlag(MODULE.ID, FLAGS.PREVIOUS_CANTRIP_MAX) || 0;
-    const currentLevel = this.actor.system.details.level;
-    const currentMax = this.getMaxAllowed();
-    return (previousLevel === 0 && currentLevel > 0) || ((currentLevel > previousLevel || currentMax > previousMax) && previousLevel > 0);
+    return this.cantripManager.canBeLeveledUp();
   }
 
   /**
@@ -597,6 +208,10 @@ export class SpellManager {
     this.isWizard = genericUtils.isWizard(this.actor);
     this._wizardSpellbookCache = null;
     this._wizardManager = null;
+
+    // Refresh sub-managers
+    this.cantripManager = new CantripManager(this.actor, this);
+    this.ritualManager = new RitualManager(this.actor);
   }
 
   /**
@@ -816,7 +431,116 @@ export class SpellManager {
    * @returns {Promise<void>}
    */
   async saveActorPreparedSpells(spellData) {
-    await saveActorPreparedSpells(this.actor, spellData);
+    try {
+      log(3, `Saving prepared spells for ${this.actor.name}`);
+      const cantripChanges = { added: [], removed: [], hasChanges: false };
+      const preparedUuids = Object.entries(spellData)
+        .filter(([_uuid, data]) => data.isPrepared)
+        .map(([uuid]) => uuid);
+
+      await this.actor.setFlag(MODULE.ID, FLAGS.PREPARED_SPELLS, preparedUuids);
+      log(3, `Saved ${preparedUuids.length} prepared spells to actor flags`);
+
+      const spellIdsToRemove = [];
+      const spellsToUpdate = [];
+      const spellsToCreate = [];
+
+      const isWizard = this.isWizard;
+      const ritualCastingEnabled = this.ritualManager.isRitualCastingEnabled();
+
+      for (const [uuid, data] of Object.entries(spellData)) {
+        if (data.isAlwaysPrepared) continue;
+        const isRitual = data.isRitual || false;
+        const existingSpell = this.actor.items.find((i) => i.type === 'spell' && (i.flags?.core?.sourceId === uuid || i.uuid === uuid));
+
+        if (!data.isPrepared) {
+          if (data.wasPrepared && existingSpell) {
+            if (isRitual && isWizard && ritualCastingEnabled) {
+              spellsToUpdate.push({
+                '_id': existingSpell.id,
+                'system.preparation.mode': 'ritual',
+                'system.preparation.prepared': false
+              });
+            } else if (existingSpell.system.preparation?.mode === 'prepared' && !existingSpell.system.preparation?.alwaysPrepared) {
+              spellIdsToRemove.push(existingSpell.id);
+              if (existingSpell.system.level === 0) {
+                cantripChanges.removed.push({ name: existingSpell.name, uuid: uuid });
+                cantripChanges.hasChanges = true;
+              }
+            }
+          } else if (isRitual && isWizard && ritualCastingEnabled && !existingSpell) {
+            try {
+              const sourceSpell = await fromUuid(uuid);
+              if (sourceSpell) {
+                const newSpellData = sourceSpell.toObject();
+                if (!newSpellData.system.preparation) newSpellData.system.preparation = {};
+                newSpellData.system.preparation.mode = 'ritual';
+                newSpellData.system.preparation.prepared = false;
+                newSpellData.flags = newSpellData.flags || {};
+                newSpellData.flags.core = newSpellData.flags.core || {};
+                newSpellData.flags.core.sourceId = uuid;
+                spellsToCreate.push(newSpellData);
+              }
+            } catch (error) {
+              log(1, `Error fetching ritual spell ${uuid}:`, error);
+            }
+          }
+        } else {
+          if (existingSpell) {
+            if (!existingSpell.system.preparation?.prepared || existingSpell.system.preparation?.mode !== 'prepared') {
+              spellsToUpdate.push({
+                '_id': existingSpell.id,
+                'system.preparation.mode': 'prepared',
+                'system.preparation.prepared': true
+              });
+            }
+          } else {
+            try {
+              const sourceSpell = await fromUuid(uuid);
+              if (sourceSpell) {
+                const newSpellData = sourceSpell.toObject();
+                if (!newSpellData.system.preparation) newSpellData.system.preparation = {};
+                newSpellData.system.preparation.mode = 'prepared';
+                newSpellData.system.preparation.prepared = true;
+                newSpellData.flags = newSpellData.flags || {};
+                newSpellData.flags.core = newSpellData.flags.core || {};
+                newSpellData.flags.core.sourceId = uuid;
+                spellsToCreate.push(newSpellData);
+
+                if (sourceSpell.system.level === 0) {
+                  cantripChanges.added.push({ name: sourceSpell.name, uuid });
+                  cantripChanges.hasChanges = true;
+                }
+              }
+            } catch (error) {
+              log(1, `Error fetching spell ${uuid}:`, error);
+            }
+          }
+        }
+      }
+
+      if (spellIdsToRemove.length > 0) {
+        log(3, `Removing ${spellIdsToRemove.length} spells from actor`);
+        await this.actor.deleteEmbeddedDocuments('Item', spellIdsToRemove);
+      }
+
+      if (spellsToUpdate.length > 0) {
+        log(3, `Updating ${spellsToUpdate.length} spells on actor`);
+        await this.actor.updateEmbeddedDocuments('Item', spellsToUpdate);
+      }
+
+      if (spellsToCreate.length > 0) {
+        log(3, `Creating ${spellsToCreate.length} spells on actor`);
+        await this.actor.createEmbeddedDocuments('Item', spellsToCreate);
+      }
+
+      if (cantripChanges.hasChanges) {
+        await this.notifyGMOfCantripChanges(cantripChanges);
+      }
+    } catch (error) {
+      log(1, `Error saving prepared spells for ${this.actor?.name || 'unknown actor'}:`, error);
+      ui.notifications.error(game.i18n.localize('SPELLBOOK.Error.SavingFailed'));
+    }
   }
 
   /**
@@ -824,12 +548,7 @@ export class SpellManager {
    * @returns {Promise<boolean>} Success status
    */
   async completeCantripsLevelUp() {
-    const currentLevel = this.actor.system.details.level;
-    const currentMax = this.getMaxAllowed();
-    await this.actor.setFlag(MODULE.ID, FLAGS.PREVIOUS_LEVEL, currentLevel);
-    await this.actor.setFlag(MODULE.ID, FLAGS.PREVIOUS_CANTRIP_MAX, currentMax);
-    await this.completeCantripSwap(true);
-    return true;
+    return this.cantripManager.completeCantripsLevelUp();
   }
 
   /**
@@ -865,14 +584,6 @@ export class SpellManager {
    * @returns {Promise<void>}
    */
   async resetSwapTracking() {
-    const allTracking = this.actor.getFlag(MODULE.ID, FLAGS.CANTRIP_SWAP_TRACKING) || {};
-
-    if (allTracking.longRest) delete allTracking.longRest;
-
-    if (Object.keys(allTracking).length === 0) {
-      await this.actor.unsetFlag(MODULE.ID, FLAGS.CANTRIP_SWAP_TRACKING);
-    } else {
-      await this.actor.setFlag(MODULE.ID, FLAGS.CANTRIP_SWAP_TRACKING, allTracking);
-    }
+    return this.cantripManager.resetSwapTracking();
   }
 }
